Issue/1388 slices (#1395)
Split out server in multiple components 

fixes #1388
