+++ src/inmanta/ast/type.py
@@ -17,7 +17,7 @@
 """
 
 import numbers
-from typing import Optional
+from typing import Any, Optional
 
 from inmanta.ast import DuplicateException, Locatable, Location, Named, Namespace, RuntimeException, TypeNotFoundException
 from inmanta.execute.util import AnyType, NoneValue
@@ -100,7 +100,7 @@ class Type(Locatable):
         raise NotImplementedError()
 
     @classmethod
-    def cast(cls, value):
+    def cast(cls, value: Any) -> Any:
         """
             Cast the given value to this type. If this fails a CastException
             is thrown.
@@ -187,7 +187,7 @@ class Number(Type):
         return True  # allow this function to be called from a lambda function
 
     @classmethod
-    def cast(cls, value):
+    def cast(cls, value: Any) -> int:
         """
             Cast the value to a number.
 
@@ -249,7 +249,7 @@ class Bool(Type):
             raise RuntimeException(None, "Invalid value '%s', expected Bool" % value)
 
     @classmethod
-    def cast(cls, value):
+    def cast(cls, value: Any) -> bool:
         """
             Convert the given value to value that can be used by the operators
             defined on this type.
@@ -285,7 +285,7 @@ class String(Type, str):
         str.__init__(self)
 
     @classmethod
-    def cast(cls, value):
+    def cast(cls, value: Any) -> str:
         """
             Cast the given value to a string
 
@@ -497,7 +497,7 @@ class ConstraintType(NamedType):
 
     constraint = property(get_constaint, set_constraint)
 
-    def cast(self, value):
+    def cast(self, value: Any) -> str:
         """
             Cast the value to the basetype of this constraint
         """
