+++ src/inmanta/config.py
@@ -290,7 +290,7 @@ class Option(Generic[T]):
     def validate(self, value: str) -> T:
         return self.validator(value)
 
-    def get_default_value(self):
+    def get_default_value(self) -> Union[str, bool, int, None]:
         defa = self.default
         if callable(defa):
             return defa()
@@ -327,10 +327,11 @@ log_dir = Option(
     "log_dir",
     "/var/log/inmanta",
     "The directory where the resource action log is stored and the logs of auto-started agents.",
+    is_str,
 )
 
 
-def get_executable():
+def get_executable() -> Optional[str]:
     """``os.path.abspath(sys.argv[0])``"""
     try:
         return os.path.abspath(sys.argv[0])
@@ -338,7 +339,7 @@ def get_executable():
         return None
 
 
-def get_default_nodename():
+def get_default_nodename() -> str:
     """``socket.gethostname()``"""
     import socket
