+++ src/inmanta/server/services/__init__.py
@@ -1,5 +1,5 @@
 """
-    Copyright 2018 Inmanta
+    Copyright 2019 Inmanta
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -15,22 +15,3 @@
 
     Contact: code@inmanta.com
 """
-
-# flake8: noqa: F401
-
-SLICE_SERVER = "core.server"
-SLICE_AGENT_MANAGER = "core.agentmanager"
-SLICE_SESSION_MANAGER = "core.session"
-SLICE_DATABASE = "core.database"
-SLICE_TRANSPORT = "core.transport"
-SLICE_COMPILER = "core.compiler"
-SLICE_FORM = "core.forms"
-SLICE_PROJECT = "core.project"
-SLICE_ENVIRONMENT = "core.environment"
-SLICE_FILE = "core.file"
-SLICE_CODE = "core.code"
-SLICE_METRICS = "core.metrics"
-SLICE_PARAM = "core.parameters"
-SLICE_RESOURCE = "core.resource"
-SLICE_ORCHESTRATION = "core.orchestration"
-SLICE_DRYRUN = "core.dryrun"
